+++ scripts/tests/export.sh
@@ -173,3 +173,18 @@ if docker inspect earthly-export-test-6:test_linux_arm64 >/dev/null 2>&1 ; then
     echo "Expected failure"
     exit 1
 fi
+
+# Test 7: remote cache on target with only BUILDs
+echo ==== Running test 7 ====
+mkdir /tmp/earthly-export-test-7
+cd /tmp/earthly-export-test-7
+cat >> Earthfile <<EOF
+VERSION 0.6
+test7:
+    BUILD +b
+b:
+    FROM busybox:latest
+EOF
+
+# This simply tests that this does not hang (#1945).
+timeout -k 11m 10m "$earthly" --ci --push --remote-cache earthly/test-cache:export-test-7 +test7